User interface: Misra coding standards(partially) #refactor

+++ src/UserInterface/Components/InfoPanel/StepCounter.cpp
@@ -21,7 +21,7 @@ void StepCounter::render() {
 			stepCount
 		);
 
-		TTGOClass::getWatch()->tft->drawString(steps, POS_X + 46, POS_Y + 1);
+		TTGOClass::getWatch()->tft->drawString(steps, (int32_t)POS_X + 46, (int32_t)POS_Y + 1);
 		this->prevStepCount = stepCount;
 	}
 }
@@ -29,17 +29,17 @@ void StepCounter::render() {
 
 void StepCounter::clearDisplay() {
 	TTGOClass::getWatch()->tft->fillRect(
-	POS_X + 44,
-	POS_Y,
-	100,
-	TTGOClass::getWatch()->tft->fontHeight(),
-	TFT_BLACK
+		(int32_t)POS_X + 44,
+		(int32_t)POS_Y,
+		100,
+		TTGOClass::getWatch()->tft->fontHeight(),
+		TFT_BLACK
 	);
 }
 
 void StepCounter::renderIcon() {
-	TTGOClass::getWatch()->tft->fillEllipse(POS_X + 24, POS_Y + 10, 14, 8, TFT_DARKGREEN);
-	TTGOClass::getWatch()->tft->fillCircle(POS_X + 4, POS_Y + 10, 8, TFT_DARKGREEN);
-	TTGOClass::getWatch()->tft->fillRect(POS_X + 6, POS_Y + 2, 2, 18, TFT_DARKGREEN);
-	TTGOClass::getWatch()->tft->fillRect(POS_X + 10, POS_Y, 3, 16, TFT_BLACK);
+	TTGOClass::getWatch()->tft->fillEllipse((int32_t)POS_X + 24, (int32_t)POS_Y + 10, 14, 8, TFT_DARKGREEN);
+	TTGOClass::getWatch()->tft->fillCircle((int32_t)POS_X + 4, (int32_t)POS_Y + 10, 8, TFT_DARKGREEN);
+	TTGOClass::getWatch()->tft->fillRect((int32_t)POS_X + 6, (int32_t)POS_Y + 2, 2, 18, TFT_DARKGREEN);
+	TTGOClass::getWatch()->tft->fillRect((int32_t)POS_X + 10, (int32_t)POS_Y, 3, 16, TFT_BLACK);
 }
